added better log msgs
