+++ slashing-protection/src/integration-test/java/tech/pegasys/web3signer/slashingprotection/InterchangeImportConflicts.java
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2020 ConsenSys AG.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+package tech.pegasys.web3signer.slashingprotection;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import com.google.common.io.Resources;
+import java.io.IOException;
+import java.net.URL;
+import org.junit.jupiter.api.Test;
+
+public class InterchangeImportConflicts extends InterchangeBaseIntegrationTest {
+
+
+  @Test
+  void canImportFileWhichContainsConflictingBlockAndBothInDatabase() {
+
+  }
+
+  @Test
+  void duplicateEntriesAreNotInsertedToDatabase() throws IOException {
+    final URL importFile = Resources.getResource("interchange/singleValidBlock.json");
+    slashingProtection.importData(importFile.openStream());
+    slashingProtection.importData(importFile.openStream());
+    jdbi.useHandle(handle -> assertThat(findAllBlocks(handle)).hasSize(1));
+  }
+
+
+}
