+++ slashing-protection/src/integration-test/resources/interchange/singleValidBlock.json
@@ -1,0 +1,18 @@
+{
+  "metadata": {
+    "interchange_format_version": "5",
+    "genesis_validators_root": "0x123456"
+  },
+  "data": [
+    {
+      "pubkey": "0xb845089a1457f811bfc000588fbb4e713669be8ce060ea6be3c6ece09afc3794106c91ca73acda5e5457122d58723bed",
+      "signed_blocks": [
+        {
+          "slot": "12345",
+          "signing_root": "0x4ff6f743a43f3b4f95350831aeaf0a122a1a392922c45d804280284a69eb850b"
+        }
+      ],
+      "signed_attestations": []
+    }
+  ]
+}
