+++ slashing-protection/src/main/java/tech/pegasys/web3signer/slashingprotection/interchange/InterchangeV5Importer.java
@@ -34,6 +34,7 @@ import org.apache.logging.log4j.Logger;
 import org.apache.tuweni.bytes.Bytes;
 import org.jdbi.v3.core.Handle;
 import org.jdbi.v3.core.Jdbi;
+import tech.pegasys.web3signer.slashingprotection.validator.BlockValidator;
 
 public class InterchangeV5Importer {
 
@@ -109,10 +110,21 @@ public class InterchangeV5Importer {
       throws JsonProcessingException {
     for (int i = 0; i < signedBlocksNode.size(); i++) {
       final SignedBlock jsonBlock = mapper.treeToValue(signedBlocksNode.get(i), SignedBlock.class);
-      signedBlocksDao.insertBlockProposal(
-          h,
-          new tech.pegasys.web3signer.slashingprotection.dao.SignedBlock(
-              validator.getId(), jsonBlock.getSlot(), jsonBlock.getSigningRoot()));
+      final BlockValidator blockValidator = new BlockValidator(
+          h, jsonBlock.getSigningRoot(), jsonBlock.getSlot(), validator.getId(), signedBlocksDao);
+
+      if (blockValidator.existsInDatabase()) {
+        LOG.debug("Block {} for validator {} already exists in database, not imported", i,
+            validator.getPublicKey());
+      } else if (blockValidator.directlyConflictsWithExistingEntry()) {
+        LOG.debug("Block {} for validator {} conflicts with on slot {} in database", i,
+            validator.getPublicKey(), jsonBlock.getSlot());
+      } else {
+        signedBlocksDao.insertBlockProposal(
+            h,
+            new tech.pegasys.web3signer.slashingprotection.dao.SignedBlock(
+                validator.getId(), jsonBlock.getSlot(), jsonBlock.getSigningRoot()));
+      }
     }
   }
